+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Fix issue [#62](https://github.com/intersystems/language-server/issues/62): DocumentSymbol doesn't cover full range of multi-line macro definition
 - Fix issue [#63](https://github.com/intersystems/language-server/issues/63): Don't provide completion suggestions inside a comment
 - Fix issue [#64](https://github.com/intersystems/language-server/issues/64): Code completion needs to understand that %&lt;classname&gt; is shorthand for %Library.&lt;classname&gt;
+- Fix issue [#66](https://github.com/intersystems/language-server/issues/66): Suggestions for $functions and variables should follow system.case settings
 - Fix issue [#69](https://github.com/intersystems/language-server/issues/69): Update class definition regex to support unicode characters
 
 ## [1.0.4] - 2020-10-28